Conflicts:
	competence-gui/competence-webapp/src/uzuzjmd/competence/gui/client/competenceSelection/CompetenceSelectionPanel.java

Former-commit-id: 6c46088effb4800452819cde619cd6d2b389b2e2

+++ competence-database/src/main/scala/uzuzjmd/competence/owl/dao/CompetenceOntologyDao.scala
@@ -17,12 +17,18 @@ abstract class CompetenceOntologyDao(comp: CompOntologyManager, compOntClass: Co
   }
 
   def delete {
-    val individual = persist()
-    individual.remove()
-    deleteMore()
+    if (exists) {
+      val individual = comp.getM().getIndividual(identifier)
+      individual.remove()
+      deleteMore()
+    }
   }
 
   protected def deleteMore()
   protected def persistMore()
+
+  def exists(): Boolean = {    
+    return comp.getM().getIndividual(identifier) != null
+  }
 
 }
